move new work into metaci
